+++ packages/service/src/app/[locale]/about/page.tsx
@@ -1,10 +1,5 @@
 import { AboutPage } from '@/page/about';
-import { AboutContents } from '@/widget/about/content';
 
 export default function Page() {
-  return (
-    <AboutPage>
-      <AboutContents />
-    </AboutPage>
-  );
+  return <AboutPage />;
 }
