+++ packages/service/src/page/about/index.tsx
@@ -3,14 +3,15 @@
 import { useRef } from 'react';
 import clsx from 'clsx';
 
-import { AboutToNextPage } from '@/widget/about/to-next-page';
-
 import { suite } from '@moeasy/storybook/font';
 import { MainFooter } from '@moeasy/storybook/ui/footer';
 
+import { AboutContents } from './content';
+import { AboutToNextPage } from './to-next-page';
+
 import * as styles from './about.css';
 
-export function AboutPage({ children }: { children?: React.ReactNode }) {
+export function AboutPageWrapper({ children }: { children?: React.ReactNode }) {
   const parentRef = useRef<HTMLElement>(null);
   return (
     <>
@@ -20,5 +21,13 @@ export function AboutPage({ children }: { children?: React.ReactNode }) {
         <MainFooter />
       </main>
     </>
+  );
+}
+
+export function AboutPage() {
+  return (
+    <AboutPageWrapper>
+      <AboutContents />
+    </AboutPageWrapper>
   );
 }
