+++ packages/service/src/widget/main/fallback.css.ts
@@ -3,7 +3,7 @@ import { globalVars } from '@moeasy/storybook/utils/styles/global.css';
 
 import { style } from '@vanilla-extract/css';
 
-export const sectionHeaderFallaback = style({
+export const sectionHeaderFallback = style({
   width: rem(230),
   height: rem(35),
   backgroundColor: globalVars.color.neutral[5],
