+++ packages/service/src/widget/main/shared/card/common.tsx
