+++ packages/service/src/widget/main/shared/card/schedule/fallback.tsx
