+++ packages/service/src/widget/main/shared/card/schedule/schedule.css.ts
